// src/rules/spells/yantra/Yantra.type.ts
export enum YantraType {
  demesne = 'demesne',
  location = 'location',
  verge = 'verge',
  roteSkill = 'roteSkill',
  concentration = 'concentration',
  highSpeech = 'highSpeech',
  runes = 'runes',
  dedicatedTool = 'dedicatedTool',
  pathTool = 'pathTool',
  orderTool = 'orderTool',
  materialSympathy = 'materialSympathy',
  representationalSympathy = 'representationalSympathy',
  symbolicSympathy = 'symbolicSympathy',
  sacrament = 'sacrament',
  rareSacrament = 'rareSacrament',
  otherworldlySacrament = 'otherworldlySacrament',
  persona = 'persona',
  custom = 'custom',
}